dt/telegram/config/bot_config.py
import os

from dotenv import load_dotenv

load_dotenv()
TOKEN = os.getenv("BOT_TELEGRAM_TOKEN")
POSTGRES_USER = os.getenv("POSTGRES_USER")
POSTGRES_PASSWORD = os.getenv("POSTGRES_PASSWORD")
POSTGRES_DB = os.getenv("POSTGRES_DB")
POSTGRES_HOST = os.getenv("POSTGRES_HOST")
